+++ lib/db/base/db_entry.dart
